requirements.txt
pyrogram==2.0.106
aiohttp
aiosqlite==0.19.0
python-dotenv==1.0.0
APScheduler==3.10.4
python-whois==0.7.3
